add MobileTicker
